split out options getter
