+++ src/background.ts
@@ -1,40 +1,15 @@
-import { GET_LOCAL_STORAGE, initialOptions, initialOptionsBool, isTrue, OptionsBool, OPTION_KEYS } from './constants';
+import { GET_LOCAL_STORAGE } from './constants';
+import { getOptions } from './options';
 import { MessageRequest, MessageResponseBool } from './utils';
 
 // バックグラウンドで実行される
-
-const MIGRATED_TO_CHROME_STORAGE = 'MIGRATED_TO_CHROME_STORAGE';
 
 window.chrome.runtime.onMessage.addListener(
   (request: MessageRequest, _, sendResponse: (res: MessageResponseBool) => void) => {
     // console.log(chrome.runtime.lastError);
     if (request.method === GET_LOCAL_STORAGE) {
-      // chrome.storageから取ってきつつ,
-      // まだ移行してないときはlocalStorageあるいは初期値を移行
-      chrome.storage.sync.get(MIGRATED_TO_CHROME_STORAGE, (isMigrated) => {
-        if (!isMigrated[MIGRATED_TO_CHROME_STORAGE]) {
-          const newOptions = { ...initialOptions, ...localStorage };
-          // 真偽値にして移行する
-          const newOptionsBool = { ...initialOptionsBool };
-          OPTION_KEYS.forEach((key) => {
-            newOptionsBool[key] = newOptions[key] === isTrue;
-          });
-          chrome.storage.sync.set(
-            {
-              ...newOptionsBool,
-              [MIGRATED_TO_CHROME_STORAGE]: true,
-            },
-            () => {
-              // 移行できたら新しい値を返す
-              sendResponse({ data: newOptionsBool });
-            },
-          );
-        } else {
-          chrome.storage.sync.get(OPTION_KEYS, (got) => {
-            // 初期値をフォールバックとしておく
-            sendResponse({ data: { ...initialOptionsBool, ...got } });
-          });
-        }
+      getOptions().then((options) => {
+        sendResponse({ data: options });
       });
     } else {
       sendResponse({ data: null });
